+++ packages/jaspr/CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unreleased patch
+
+- Added `jaspr daemon` command to run a daemon server (used by the new [VSCode extension]()).
+- Added `--launch-in-chrome` option to `jaspr serve` command to open the browser automatically.
+- Added support for running `jaspr create .` in an empty directory.
+
 ## 0.18.1
 
 - Better report errors during static build.